refactoring all the codes and add method on transactions

+++ src/main/java/sg/edu/ntu/javaproject/controller/AccountController.java
@@ -4,7 +4,8 @@ import java.util.ArrayList;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import sg.edu.ntu.javaproject.entity.Account;
 import sg.edu.ntu.javaproject.service.AccountService;
@@ -33,13 +35,22 @@ public class AccountController {
         this.objectMapper = objectMapper;
     }
 
+    private boolean isAdmin() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ADMIN"));
+    }
+
     @PostMapping({ "", "/" })
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Account> createAccount(@RequestBody Account account) throws JsonProcessingException {
-        Account newAccount = accountService.createAccount(account);
-        String accountJson = objectMapper.writeValueAsString(newAccount);
-        log.info("new account created: " + accountJson);
-        return new ResponseEntity<>(newAccount, HttpStatus.CREATED);
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<?> createAccount(@Valid @RequestBody Account account) throws JsonProcessingException {
+        if (isAdmin()) {
+            Account newAccount = accountService.createAccount(account);
+            String accountJson = objectMapper.writeValueAsString(newAccount);
+            log.info("new account created: " + accountJson);
+            return new ResponseEntity<>(newAccount, HttpStatus.CREATED);
+        } else
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("access denied!");
     }
 
     @GetMapping({ "", "/" })
@@ -52,40 +63,53 @@ public class AccountController {
     }
 
     @GetMapping({ "/{id}", "/{id}/" })
-    public ResponseEntity<Account> getAccountById(@PathVariable Integer id) throws JsonProcessingException {
-        Account accountById = accountService.getAccountById(id);
-        String accountJson = objectMapper.writeValueAsString(accountById);
-        log.info("search account by account id " + id);
-        log.info("account details: " + accountJson);
-        return new ResponseEntity<>(accountById, HttpStatus.OK);
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<?> getAccountById(@PathVariable Integer id) throws JsonProcessingException {
+        if (isAdmin()) {
+            Account accountById = accountService.getAccountById(id);
+            String accountJson = objectMapper.writeValueAsString(accountById);
+            log.info("search account by account id " + id);
+            log.info("account details: " + accountJson);
+            return new ResponseEntity<>(accountById, HttpStatus.OK);
+        } else
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("access denied!");
     }
 
-    // Lionel: Hendry may I ask u if we are searching for an account by the user id
-    // or are we searching for an account id by the user?
     @GetMapping("/searchByCustomerId/{id}")
-    public ResponseEntity<ArrayList<Account>> searchByCustomerId(@PathVariable Integer id)
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<?> searchByCustomerId(@PathVariable Integer id)
             throws JsonProcessingException {
-        ArrayList<Account> accountList = accountService.getAccountByCustomerId(id);
-        String accountJson = objectMapper.writeValueAsString(accountList);
-        log.info("search accounts   by customer id: " + id);
-        log.info("account list: " + accountJson);
-        return new ResponseEntity<>(accountList, HttpStatus.OK);
+        if (isAdmin()) {
+            ArrayList<Account> accountList = accountService.getAccountByCustomerId(id);
+            String accountJson = objectMapper.writeValueAsString(accountList);
+            log.info("search accounts   by customer id: " + id);
+            log.info("account list: " + accountJson);
+            return new ResponseEntity<>(accountList, HttpStatus.OK);
+        } else
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("access denied!");
     }
 
     @DeleteMapping({ "/{id}", "/{id}/" })
-    public ResponseEntity<Account> deleteAccountById(@PathVariable Integer id) {
-        accountService.deleteAccountById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<?> deleteAccountById(@PathVariable Integer id) {
+        if (isAdmin()) {
+            accountService.deleteAccountById(id);
+            return ResponseEntity.status(HttpStatus.OK).body("account id: " + id + " has been deleted");
+        } else
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("access denied!");
     }
 
     @PutMapping({ "/{id}", "/{id}/" })
-    public ResponseEntity<Account> updateAccountById(@PathVariable Integer id, @RequestBody Account account)
+    // @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<?> updateAccountById(@PathVariable Integer id, @RequestBody Account account)
             throws JsonProcessingException {
-        Account updatedAccount = accountService.updateAccount(id, account);
-        String accountJson = objectMapper.writeValueAsString(updatedAccount);
-        log.info("updating account id: " + id);
-        log.info("new account details: " + accountJson);
-        return new ResponseEntity<>(updatedAccount, HttpStatus.OK);
+        if (isAdmin()) {
+            Account updatedAccount = accountService.updateAccount(id, account);
+            String accountJson = objectMapper.writeValueAsString(updatedAccount);
+            log.info("updating account id: " + id);
+            log.info("new account details: " + accountJson);
+            return new ResponseEntity<>(updatedAccount, HttpStatus.OK);
+        } else
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("access denied!");
     }
-
 }
